added (int) cast to uint8_t before subtracting them from each other in calculating of cost...

+++ sw/airborne/modules/computer_vision/lib/vision/qr_code_recognition.c
@@ -175,7 +175,7 @@ void get_closest_template(struct image_t *img, uint32_t x_min, uint32_t y_min, u
         Y_template = get_sub_pixel(img, sub_pixel_x, sub_pixel_y, templ);
 
         // augment the SAD cost:
-        cost += abs(Y_im - Y_template);
+        cost += abs((int)Y_im - (int)Y_template);
       }
     }
